+++ src/Defaults.test.ts
@@ -1,4 +1,4 @@
-describe('module test', () => {
+describe('Defaults test', () => {
   it.skip('plugin should be defined', () => {
     expect(true).toBeTruthy();
   });