+++ src/Utils.test.ts
@@ -1,0 +1,5 @@
+describe('Utils test', () => {
+  it.skip('GetColorForValue', () => {
+    expect(true).toBeTruthy();
+  });
+});